+++ src/app/[orderCode]/OrderView.tsx
@@ -18,9 +18,9 @@ import {
   Stack,
   CircularProgress,
 } from "@mui/material";
-import Layout from "../../../components/layout";
-import FormDialog from "../../../components/form-dialog";
-import FormEditDialog from "../../../components/form-edit-dialog";
+import Layout from "../../components/Layout";
+import FormDialog from "../../components/form-dialog";
+import FormEditDialog from "../../components/form-edit-dialog";
 import OrderItemService from "../../../services/OrderItemService";
 import { useMenuItems } from "@/hooks/useMenuItems";
