+++ src/app/page.tsx
@@ -2,8 +2,8 @@ import Link from "next/link";
 import { Typography } from "@mui/material";
 import { useTranslations } from 'next-intl';
 import styles from "@/styles/Home.module.css";
-import Footer from "@/components/footer";
-import UserWelcome from "@/components/user-welcome";
+import Footer from "@/src/components/Footer";
+import UserWelcome from "@/src/components/UserWelcome";
 
 
 export default function Home() {
